// web/src/api/generated/schemas/worksheetResource.zod.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Fonzytooter API
 * OpenAPI spec version: 0.1.0
 */
import * as zod from 'zod'
import { WorksheetProblemResource } from './worksheetProblemResource.zod'

export const WorksheetResource = zod.strictObject({
  $schema: zod.url().optional().describe('A URL to the JSON Schema for this object.'),
  courseId: zod.string(),
  id: zod.string(),
  instructions: zod.string(),
  lessonId: zod.string(),
  moduleId: zod.string(),
  objectiveIds: zod.array(zod.string()),
  order: zod.int(),
  problems: zod.array(WorksheetProblemResource),
  title: zod.string(),
})

export type WorksheetResource = zod.input<typeof WorksheetResource>
export type WorksheetResourceOutput = zod.output<typeof WorksheetResource>